01_21_2009 PC
git-svn-id: http://dim3.googlecode.com/svn/trunk@381 2c4a6428-2a2a-0410-ad13-7dde452bfe2e

+++ dim3Engine/Sources_View/gl_shader.c
@@ -121,7 +121,7 @@ void gl_shader_set_variables(GLhandleARB shader_prog_obj,d3pnt *pnt,texture_type
 			// camera is always at 0,0,0
 			
 		var=glGetUniformLocationARB(shader_prog_obj,"dim3CameraPosition");
-		if (var!=-1) glUniform3fARB(var,(float)view.camera.pnt.x,view.camera.pnt.y,view.camera.pnt.z);
+		if (var!=-1) glUniform3fARB(var,(float)view.camera.pnt.x,(float)view.camera.pnt.y,(float)view.camera.pnt.z);
 
 			// ambient light
